Bunny Retreat

+++ Assets/Script/Enemies/Beheaviour Tree/CheckIfPlayerInRange.cs
@@ -11,9 +11,9 @@ public class CheckIfPlayerInRange : Node
     private GameObject _player;
     private Transform _myPosition;
 
-    private int distanceToCheck;
+    private float distanceToCheck;
     
-    public CheckIfPlayerInRange(GameObject player, Transform myPosition, int distanceToCheck)
+    public CheckIfPlayerInRange(GameObject player, Transform myPosition, float distanceToCheck)
     {
         _player = player;
         _myPosition = myPosition;
@@ -23,7 +23,7 @@ public class CheckIfPlayerInRange : Node
     public override NodeState Evaluate()
     {
         float distanceToPlayer = Vector3.Distance(_player.transform.position, _myPosition.position);
-        
+
         if (distanceToPlayer < distanceToCheck)
         {
             return NodeState.SUCCESS;
